Attempt at removing call to `ps` in linux systems for metrics
This means involves relying on procfs to get the cpu and mem metrics
which is not completely portable (e.g. won't work in OS X) so original
method is still included with a conditional build tag.

Note that current implementation supposes `sysconf(_SC_CLK_TCK)`
value is 100.0, even though this is not guaranteed.