Add support for handling gnatsd config as JSON
The gnatsd config format is almost already valid JSON
with exception of treating brackets at the top level
of a block and trailing commas.

In this commit rather than changing the config parser to
relax those items, we peek at the first character to check
if it is a bracket hinting that it could be valid JSON.

When using JSON as the configuration format, a lot of the
syntax sugar from the config format is disabled so features
such as include directives and variable expansions do not
work and have to declared explicitly instead.