[FIXED] Corruption of subjects on recover of bad or missing index.db (#4851)
When populating the block's fss during a manual recovery, we could hoist
the fss to the global PSIM.
This could lead to data corruption since we were not copying. This
revert's that change with a test.

The corruption was all internal at the PSIM layer and was not reflected
on disk, so a server upgrade will fix with no data corruption.

Resolves: #4842

Signed-off-by: Derek Collison <derek@nats.io>